Initial commit
This commit contains the source code for the shopping basket lab from
the TDD 1 course.  It is at the position where we get a failing test
due to the 72.8875 rounding error that is the insight into the need of
a separate money class.

This is probably a good moment to reflect on one of the focal points
of the TDD 1 course.