Represent cost with a new class, Money
This commit will replace the number representation of cost with a new
class, Money.

Class Money will correctly handle the appropriate number of decimal
places and rounding for the Swedish currency.

All tests now pass, but note that test "should total with VAT" now
asserts the total to equal 72.89, instead of 72.88.  The value 72.88
from the previous demo was a pedagogical exaggeration, to better drive
home an important point of the course: good tests are a source of
domain insights.  You get to 72.88 by truncating the number of decimal
points to two.  Class Money will instead round to the nearest value
(72.89) since this is what Skatteverket does.